+++ lib/laddu_seva/main.dart
@@ -5,8 +5,10 @@ import 'package:garuda/laddu_seva/datatypes.dart';
 import 'package:garuda/laddu_seva/history.dart';
 import 'package:garuda/laddu_seva/laddu_calc.dart';
 import 'package:garuda/laddu_seva/log.dart';
-import 'package:garuda/laddu_seva/serve.dart';
+import 'package:garuda/laddu_seva/service_select.dart';
 import 'package:garuda/laddu_seva/summary.dart';
+import 'package:garuda/laddu_seva/utils.dart';
+import 'package:garuda/toaster.dart';
 import 'package:intl/intl.dart';
 
 class LadduMain extends StatefulWidget {
@@ -17,27 +19,22 @@ class LadduMain extends StatefulWidget {
 class _LadduSevaState extends State<LadduMain> {
   DateTime? session;
   LadduReturn? lr;
-  DateTime? lastRefresh;
 
   @override
   initState() {
     super.initState();
 
-    refresh();
+    refresh().then((data) async {
+      await _ensureReturn(context);
 
-    FB().listenForChange("ladduSeva",
-        FBCallbacks(onChange: (String changeType, dynamic data) async {
-      await refresh();
-    }));
+      FB().listenForChange("ladduSeva",
+          FBCallbacks(onChange: (String changeType, dynamic data) async {
+        await refresh();
+      }));
+    });
   }
 
   Future<void> refresh() async {
-    if (lastRefresh != null &&
-        DateTime.now().difference(lastRefresh!).inSeconds < 2) {
-      return;
-    }
-    lastRefresh = DateTime.now();
-
     // refresh the main widget
     session = await FB().readLatestLadduSession();
     lr = await FB().readLadduReturnStatus(session!);
@@ -124,6 +121,58 @@ class _LadduSevaState extends State<LadduMain> {
         });
   }
 
+  Future<void> _ensureReturn(BuildContext context) async {
+    if (lr == null || lr!.count == -1) {
+      // session in progress
+
+      DateTime session = await FB().readLatestLadduSession();
+      List<LadduServe> serves = await FB().readLadduServes(session);
+
+      // check if last serve is more than 2 days old
+      if (serves.last.timestamp
+          .isBefore(DateTime.now().subtract(Duration(days: 2)))) {
+        // totatl stock
+        List<LadduStock> stocks = await FB().readLadduStocks(session);
+        stocks.sort((a, b) => a.timestamp.compareTo(b.timestamp));
+        if (stocks.isEmpty) {
+          return;
+        }
+        int totalStock = stocks.fold(
+            0, (previousValue, element) => previousValue + element.count);
+
+        // total serve
+        int totalServe = 0;
+        serves.forEach((serve) {
+          totalServe += CalculateTotalLadduPacksServed(serve);
+        });
+
+        int remaining = totalStock - totalServe;
+        if (remaining < 0) {
+          remaining = 0;
+        }
+
+        await FB().returnLadduStock(
+            session,
+            LadduReturn(
+                timestamp: DateTime.now(),
+                count: remaining,
+                to: "Unknown",
+                user: "Auto Return"));
+
+        Toaster().info("Auto returned");
+      }
+    }
+  }
+
+  void _createServeDialog(BuildContext context) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return ServiceSelect();
+      },
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -171,11 +220,8 @@ class _LadduSevaState extends State<LadduMain> {
                 // serve button
                 ElevatedButton.icon(
                   onPressed: (lr == null || lr!.count == -1)
-                      ? () {
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(builder: (context) => Serve()),
-                          );
+                      ? () async {
+                          _createServeDialog(context);
                         }
                       : null,
                   icon: Icon(Icons.remove),