+++ lib/pushpanjali/record.dart
@@ -4,11 +4,9 @@ import 'package:garuda/pushpanjali/sevaslot.dart';
 
 class Record {
   static final Record _instance = Record._internal();
-  List<SevaSlot> sevaSlots = [];
+  List<PushpanjaliSlot> sevaSlots = [];
   Map<DateTime, List<SevaTicket>> sevaTickets = {}; // to FB from List to Map
   RecordCallbacks? callbacks;
-
-  DateTime? _lastRefresh;
 
   // Private constructor
   Record._internal();
@@ -25,15 +23,10 @@ class Record {
   }
 
   void onSevaSlotChange(String changeType, dynamic sevaSlot) {
-    if (_lastRefresh != null &&
-        DateTime.now().difference(_lastRefresh!) < const Duration(seconds: 2)) {
-      return;
-    }
-    _lastRefresh = DateTime.now();
-
     switch (changeType) {
       case 'ADD_SEVA_SLOT':
-        SevaSlot slot = SevaSlot.fromJson(Map<String, dynamic>.from(sevaSlot));
+        PushpanjaliSlot slot =
+            PushpanjaliSlot.fromJson(Map<String, dynamic>.from(sevaSlot));
 
         if (sevaSlots
             .any((element) => element.timestampSlot == slot.timestampSlot)) {
@@ -45,7 +38,8 @@ class Record {
 
         break;
       case 'REMOVE_SEVA_SLOT':
-        SevaSlot slot = SevaSlot.fromJson(Map<String, dynamic>.from(sevaSlot));
+        PushpanjaliSlot slot =
+            PushpanjaliSlot.fromJson(Map<String, dynamic>.from(sevaSlot));
         deleteSevaSlot(slot.timestampSlot);
         break;
       case 'UPDATE_SEVA_SLOT':
@@ -56,12 +50,6 @@ class Record {
   }
 
   void onSevaTicketChange(String changeType, dynamic sevaTicketMap) {
-    if (_lastRefresh != null &&
-        DateTime.now().difference(_lastRefresh!) < const Duration(seconds: 2)) {
-      return;
-    }
-    _lastRefresh = DateTime.now();
-
     switch (changeType) {
       case 'ADD_SEVA_TICKET':
         sevaTicketMap.forEach((dynamic timestampTicket, dynamic ticketData) {
@@ -122,7 +110,7 @@ class Record {
     }
   }
 
-  void addSevaSlot(DateTime timestampSlot, SevaSlot slot) {
+  void addSevaSlot(DateTime timestampSlot, PushpanjaliSlot slot) {
     sevaSlots.add(slot);
     sevaSlots.sort((a, b) => b.timestampSlot.compareTo(a.timestampSlot));
 
@@ -214,12 +202,13 @@ class Record {
   }
 
   Future<void> refreshSevaSlots() async {
-    sevaSlots = await FB().readSevaSlots();
+    sevaSlots = await FB().readPushpanjaliSlots();
     sevaSlots.sort((a, b) => b.timestampSlot.compareTo(a.timestampSlot));
   }
 
   Future<void> refreshSevaTickets(DateTime timestampSlot) async {
-    sevaTickets[timestampSlot] = await FB().readSevaTickets(timestampSlot);
+    sevaTickets[timestampSlot] =
+        await FB().readPushpanjaliTickets(timestampSlot);
     sevaTickets[timestampSlot]!
         .sort((a, b) => b.timestampTicket.compareTo(a.timestampTicket));
   }